requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn>=1.3.0
nltk==3.8.1
numpy>=1.24.0
gunicorn==21.2.0
Werkzeug>=3.0.0
